Create .babelrc on test ensure that the JavaScript tests run correctly
When using babel with mocha, a .babelrc file is required to ensure that
the tests are transpiled. Adding .babelrc in the Phoenix application
causes issues when Phoenix is used as a dependency (via git or file
path) due to the application being used as a JavaScript dependency.

To fix both of these issues, a `.babelrc` file is created when running
the tests, however it is also .gitignored to prevent it being
accidentally committed.