Use assets directory for building JS and tests
When using babel with mocha, a .babelrc file is required to ensure that
the tests are transpiled. Adding .babelrc in the Phoenix application
causes issues when Phoenix is used as a dependency (via git or file
path) due to the application being used as a JavaScript dependency.

To fix both of these issues, all building and testing that involved
JavaScript has been moved to the assets directory. This means that when
the Phoenix package is included as a node dependency, the root
`package.json` will be used, but when building, the
`assets/package.json` should be used.